requirements.txt
click >= 0.0
pyyaml >= 0.0
jinja2 >= 0.0
dacite >= 0.0
dataclasses >= 0.0
dataclasses-json >= 0.0
dataclasses-jsonschema >= 0.0
PyGithub >= 0.0